* ffmpeg_track/master: (202 commits)
  vorbisdec: Accept 0 amplitude_bits
  vorbisdec: Add missing checks
  vorbisdec: cosmetics
  vorbisdec: Error on bark_map_size equal to 0.
  af_biquads: memset(0) cache
  swr: make the default of nopts for first_pts actually work
  fate: force a first_pts=0 for the aresample test
  ffmpeg: Force a first_pts of 0 for the first configuration of -async use
  ffplay: reindent and remove pointless cast in audio_decode_frame() code
  ffplay: set type for channel_layout AudioParams field to int64_t
  lavfi/abuffersink: fix weird indent and spacing
  lavfi/abuffersink: add sample_rates field to AVABufferSinkParams
  lavc: change type of AVFrame.channels field from int64_t to int
  ffmpeg: use a AVDictionary instead of the context to move swr parameters around
  cmdutils: fix null pointer dereference
  swr: support a seperate output sample bits.
  avfilter/avfiltergraph.h: Move public field out of the private fields
  swr: add duplicate cutoff for compatibility
  fate: add animated gif decoder tests
  fate: add animated gif demuxer test
  ...

Conflicts:
	configure
	libavformat/utils.c

git-svn-id: http://svn.ad-group.adh/svn/native_client/native/ffmpeg/branches/adffmpeg-trunk-tracker@9337 4601b0d2-a6ba-4ff0-a3bb-59d57de6f651

+++ libavcodec/avcodec.h
@@ -1510,7 +1510,7 @@ typedef struct AVFrame {
      * - encoding: unused
      * - decoding: Read by user.
      */
-    int64_t channels;
+    int channels;
 
     /**
      * size of the corresponding packet containing the compressed
